requirements.txt
pyspark==3.5.0
great-expectations==0.18.8
pandas==2.0.3
PyMySQL==1.1.0
mysqlclient==2.2.1
mysql-connector-python==8.3.0
faker==22.5.1